+++ tests/Unit/AlgoliaEngineTest.php
@@ -156,17 +156,23 @@ class AlgoliaEngineTest extends TestCase
         $engine = new AlgoliaEngine($client);
 
         $model = m::mock(stdClass::class);
+
         $model->shouldReceive('getScoutModelsByIds')->andReturn($models = Collection::make([
-            new SearchableModel(['id' => 1]),
+            new SearchableModel(['id' => 1, 'name' => 'test']),
         ]));
 
         $builder = m::mock(Builder::class);
 
-        $results = $engine->map($builder, ['nbHits' => 1, 'hits' => [
-            ['objectID' => 1, 'id' => 1],
-        ]], $model);
+        $results = $engine->map($builder, [
+            'nbHits' => 1,
+            'hits' => [
+                ['objectID' => 1, 'id' => 1, '_rankingInfo' => ['nbTypos' => 0]],
+            ],
+        ], $model);
 
         $this->assertCount(1, $results);
+        $this->assertEquals(['id' => 1, 'name' => 'test'], $results->first()->toArray());
+        $this->assertEquals(['_rankingInfo' => ['nbTypos' => 0]], $results->first()->scoutMetaData());
     }
 
     public function test_map_method_respects_order()
@@ -210,16 +216,18 @@ class AlgoliaEngineTest extends TestCase
 
         $model = m::mock(stdClass::class);
         $model->shouldReceive('queryScoutModelsByIds->cursor')->andReturn($models = LazyCollection::make([
-            new SearchableModel(['id' => 1]),
+            new SearchableModel(['id' => 1, 'name' => 'test']),
         ]));
 
         $builder = m::mock(Builder::class);
 
         $results = $engine->lazyMap($builder, ['nbHits' => 1, 'hits' => [
-            ['objectID' => 1, 'id' => 1],
+            ['objectID' => 1, 'id' => 1, '_rankingInfo' => ['nbTypos' => 0]],
         ]], $model);
 
         $this->assertCount(1, $results);
+        $this->assertEquals(['id' => 1, 'name' => 'test'], $results->first()->toArray());
+        $this->assertEquals(['_rankingInfo' => ['nbTypos' => 0]], $results->first()->scoutMetaData());
     }
 
     public function test_lazy_map_method_respects_order()